+++ packages/component/src/stories/hooks/usePoster.stories.tsx
@@ -19,11 +19,11 @@ const meta: Meta = {
     lossColor: "rgb(255,103,194)",
     brandColor: "rgb(0,181,159)",
     data: {
-      message: "I am the WOO KING.",
-      domain: "dex.woo.org",
+      message: "I am the Orderly KING.",
+      domain: "ordely.network",
       updateTime: "2022-JAN-01 23:23",
       position: {
-        symbol: "WOO-PERP",
+        symbol: "BTC-PERP",
         currency: "USDC",
         side: "LONG",
         leverage: 20,