+++ packages/utils/src/index.ts
@@ -5,4 +5,4 @@ export * from "./decimal";
 export * from "./dateTime";
 export * from "./chain";
 export * from "./string";
-export { windowGuard, getGlobalObject } from "./window";
+export { windowGuard, getGlobalObject, getTimestamp } from "./window";