+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @orderly.network/futures
 
+## 3.3.4
+
+### Patch Changes
+
+- e8bc10c4: add IDs;
+- Updated dependencies [e8bc10c4]
+  - @orderly.network/types@1.5.4
+  - @orderly.network/utils@1.5.4
+
+## 3.3.4-alpha.0
+
+### Patch Changes
+
+- add IDs;
+- Updated dependencies
+  - @orderly.network/types@1.5.4-alpha.0
+  - @orderly.network/utils@1.5.4-alpha.0
+
+## 3.3.3
+
+### Patch Changes
+
+- 77c9c09b: fix bug;
+- Updated dependencies [77c9c09b]
+  - @orderly.network/utils@1.5.3
+  - @orderly.network/types@1.5.3
+
+## 3.3.3-alpha.0
+
+### Patch Changes
+
+- fix bug;
+- Updated dependencies
+  - @orderly.network/utils@1.5.3-alpha.0
+  - @orderly.network/types@1.5.3-alpha.0
+
 ## 3.3.2
 
 ### Patch Changes